+++ s/core/taps/dud.ts
@@ -3,7 +3,7 @@ import {Tap} from "../types.js"
 
 export class DudTap implements Tap {
 	error: Tap["error"] = async() => {}
-	request: Tap["request"] = async() => {}
+	rpcRequest: Tap["rpcRequest"] = async() => {}
 	rpcError: Tap["rpcError"] = async() => {}
 }