+++ s/core/taps/error.ts
@@ -3,6 +3,6 @@ import {Tap} from "../types.js"
 import {LoggerTap} from "./logger.js"
 
 export class ErrorTap extends LoggerTap implements Tap {
-	request: Tap["request"] = async() => {}
+	rpcRequest: Tap["rpcRequest"] = async() => {}
 }